+++ lib/lorina/lorina/lorina.hpp
@@ -1,5 +1,5 @@
 /* lorina: C++ parsing library
- * Copyright (C) 2021  EPFL
+ * Copyright (C) 2018-2021  EPFL
  *
  * Permission is hereby granted, free of charge, to any person
  * obtaining a copy of this software and associated documentation
@@ -36,6 +36,7 @@
 #include "bench.hpp"
 #include "blif.hpp"
 #include "bristol.hpp"
-#include "diamcs.hpp"
+#include "dimacs.hpp"
+#include "genlib.hpp"
 #include "pla.hpp"
 #include "verilog.hpp"